+++ modules/promise/index.ts
@@ -1,3 +1,5 @@
-export { DevRunnerResolver } from './dev/runner.resolver';
+export { DevRunnerResolver } from './resolvers/dev-runner.resolver';
+export { NodeRunnerResolver } from './resolvers/node-runner.resolver';
 export { RunnerResolver } from './resolvers/runner.resolver';
+export { WorkerRunnerResolver } from './resolvers/worker-runner.resolver';
