+++ modules/rx/index.ts
@@ -1,3 +1,11 @@
-export { RxDevRunnerResolver } from './dev/runner.resolver';
+export { RxNodeAction } from './actions/node.actions';
+export { RxWorkerAction } from './actions/worker.actions';
+export { RxResolveRunner, RxSerializeRunnerMethod } from './resolved-runner';
+export { RxDevRunnerResolver } from './resolvers/dev-runner.resolver';
+export { RxNodeRunnerResolver } from './resolvers/node-runner.resolver';
 export { RxRunnerResolver } from './resolvers/runner.resolver';
+export { RxWorkerRunnerResolver } from './resolvers/worker-runner.resolver';
+export { RxRunnerErrorCode, RxRunnerErrorMessages } from './runners-errors';
+export { RxNodeRunnerState } from './states/node-runner.state';
+export { RxWorkerRunnerState } from './states/worker-runner.state';
