+++ package.json
@@ -23,7 +23,7 @@
   "devDependencies": {
     "@types/jasmine": "^3.5.0",
     "@types/karma-webpack": "^2.0.7",
-    "@types/webpack": "^4.41.0",
+    "@types/webpack": "^4.41.1",
     "copy-webpack-plugin": "^5.1.1",
     "jasmine": "^3.5.0",
     "karma": "^4.4.1",
@@ -34,10 +34,10 @@
     "ts-loader": "^6.2.1",
     "tsconfig-paths-webpack-plugin": "^3.2.0",
     "tslint": "^5.20.1",
-    "typescript": "^3.7.3",
+    "typescript": "^3.7.4",
     "webpack": "^4.41.5"
   },
   "dependencies": {
-    "rxjs": "^6.5.3"
+    "rxjs": "^6.5.4"
   }
 }
