+++ test/common/promise.ts
@@ -1,5 +1,4 @@
-import { DevRunnerResolver } from '@worker-runner/promise/dev/runner.resolver';
-import { RunnerResolver } from '@worker-runner/promise/resolvers/runner.resolver';
+import { DevRunnerResolver, RunnerResolver } from '@worker-runner/promise';
 import { runners } from './runner-list';
 
 export const runnerResolver = new RunnerResolver({runners, workerPath: 'base/test/worker.js'});
