+++ test/common/rx.ts
@@ -1,5 +1,4 @@
-import { RxDevRunnerResolver } from '@worker-runner/rx/dev/runner.resolver';
-import { RxRunnerResolver } from '@worker-runner/rx/resolvers/runner.resolver';
+import { RxDevRunnerResolver, RxRunnerResolver } from '@worker-runner/rx';
 import { runners } from './runner-list';
 
 export const rxRunnerResolver = new RxRunnerResolver({runners, workerPath: 'base/test/rx-worker.js'});
